chore: add missed changes
