+++ packages/peer-discovery-mdns/test/multicast-dns.spec.ts
@@ -6,7 +6,8 @@ import { multiaddr } from '@multiformats/multiaddr'
 import { expect } from 'aegir/chai'
 import pWaitFor from 'p-wait-for'
 import { stubInterface } from 'ts-sinon'
-import { mdns, type MulticastDNSComponents } from './../src/index.js'
+import { mdns } from './../src/index.js'
+import type { MulticastDNSComponents } from './../src/mdns.js'
 import type { PeerId } from '@libp2p/interface/peer-id'
 import type { PeerInfo } from '@libp2p/interface/peer-info'
 import type { AddressManager } from '@libp2p/interface-internal/address-manager'
